Equality checks.
Apparently there's ONE FREAKING THING different in two subsequent executions
of some code that runs through both ActiveRecord and RestClient, namely
the contents of ActiveRecord::LogSubscriber#sql. REALLY?!?!

On the plus side, if it's only one thing, a thing which, moreover, could
potentially be whitelisted, then maybe checking execution paths to see if
values we come up with while testing actually change the behavior of the
program has some merit. The whole thing could moreover start ignoring pieces
that only touch external libraries if they differ on subsequent runs and warn
the user about it.

(And perhaps ignore, but warn with a much louder voice, similar issues with
the code to be tested, and just fall back to watching comparisons between
values to see what sorts of values might cause things to change.)

Now to figure out how to not consume 97.3 MB of RAM tracking an entire
invocation tree while running a test...